blindkit/requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.3.2
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
comtypes==1.4.10
contourpy==1.3.1
cycler==0.12.1
easyocr==1.7.2
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
gast==0.4.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-cloud-speech==2.31.1
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.63.0rc1
h5py==3.13.0
httplib2==0.22.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
keras==2.13.1
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
multidict==6.3.2
networkx==3.4.2
ninja==1.11.1.4
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
playsound==1.2.2
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.6
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pyclipper==1.3.0.post6
pydantic==1.10.21
PyJWT==2.10.1
pyparsing==3.2.3
pypiwin32==223
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pyttsx3==2.98
pytz==2025.2
pywin32==310
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.25.2
scipy==1.15.2
seaborn==0.13.2
shapely==2.1.0
six==1.17.0
sympy==1.13.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.0.1
tifffile==2025.3.30
torch==2.1.2
torchvision==0.16.2
tqdm==4.67.1
twilio==9.5.1
typing_extensions==4.5.0
tzdata==2025.2
ultralytics==8.3.101
ultralytics-thop==2.0.14
uritemplate==4.1.1
urllib3==2.3.0
Wave==0.0.2
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3